Removed Sequelizer and extented mysql file
Removed Sequelizer because it was not necessary; it's better for more complex purposes. Our requirements are not complex.

Extented the mysql node.js file to handle the database connection, fuel pump table creation, and querying said table.

May need to be updated to add more functionality and/or fix bugs.